// src/types/software/software.form.types.ts
export interface SoftwareFormValues {
  ministryId: string;
  softwareName: string;
  softwarePurpose: string;
  manufacturer: string;
  supplier: string;
  purchaseDate: string;
  purchaseAmount: string;
  purchaseCurrencyId: string;
  softwareVersion: string;
  lastUpdateDate: string;
  licenseType: string;
  licenseExpiryDate: string;
  licenseCount: string;
}